+++ xarray/tests/test_dataarray.py
@@ -57,30 +57,34 @@ class TestDataArray:
         data_array = DataArray(v, coords, name="my_variable")
         expected = dedent(
             """\
-        <xarray.DataArray 'my_variable' (time: 2, x: 3)>
-        array([[1, 2, 3],
-               [4, 5, 6]])
-        Coordinates:
-          * x        (x) int64 0 1 2
-            other    int64 0
-        Dimensions without coordinates: time
-        Attributes:
-            foo:      bar"""
+            <xarray.DataArray 'my_variable' (time: 2, x: 3)>
+            array([[1, 2, 3],
+                   [4, 5, 6]])
+            Coordinates:
+              * x        (x) int64 0 1 2
+                other    int64 0
+            Dimensions without coordinates: time
+            Attributes:
+                foo:      bar"""
         )
         assert expected == repr(data_array)
 
     def test_repr_multiindex(self):
         expected = dedent(
             """\
-        <xarray.DataArray (x: 4)>
-        array([0, 1, 2, 3])
-        Coordinates:
-          * x        (x) MultiIndex
-          - level_1  (x) object 'a' 'a' 'b' 'b'
-          - level_2  (x) int64 1 2 1 2"""
+            <xarray.DataArray (x: 4)>
+            array([0, 1, 2, 3])
+            Coordinates:
+              * x        (x) MultiIndex
+              - level_1  (x) object 'a' 'a' 'b' 'b'
+              - level_2  (x) int64 1 2 1 2"""
         )
         assert expected == repr(self.mda)
 
+    @pytest.mark.skipif(
+        LooseVersion(np.__version__) < "1.15",
+        reason="old versions of numpy have different printing behavior",
+    )
     def test_repr_multiindex_long(self):
         mindex_long = pd.MultiIndex.from_product(
             [["a", "b", "c", "d"], [1, 2, 3, 4, 5, 6, 7, 8]],
@@ -89,13 +93,13 @@ class TestDataArray:
         mda_long = DataArray(list(range(32)), coords={"x": mindex_long}, dims="x")
         expected = dedent(
             """\
-        <xarray.DataArray (x: 32)>
-        array([ 0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14, 15, 16, 17,
-               18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31])
-        Coordinates:
-          * x        (x) MultiIndex
-          - level_1  (x) object 'a' 'a' 'a' 'a' 'a' 'a' 'a' ... 'd' 'd' 'd' 'd' 'd' 'd'
-          - level_2  (x) int64 1 2 3 4 5 6 7 8 1 2 3 4 5 6 ... 4 5 6 7 8 1 2 3 4 5 6 7 8"""  # noqa: E501
+            <xarray.DataArray (x: 32)>
+            array([ 0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14, 15, 16,
+                   17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31])
+            Coordinates:
+              * x        (x) MultiIndex
+              - level_1  (x) object 'a' 'a' 'a' 'a' 'a' 'a' 'a' ... 'd' 'd' 'd' 'd' 'd' 'd'
+              - level_2  (x) int64 1 2 3 4 5 6 7 8 1 2 3 4 5 6 ... 4 5 6 7 8 1 2 3 4 5 6 7 8"""
         )
         assert expected == repr(mda_long)
 
@@ -1712,6 +1716,11 @@ class TestDataArray:
         with raises_regex(ValueError, "dimension mismatch"):
             array2d.set_index(x="level")
 
+        # Issue 3176: Ensure clear error message on key error.
+        with pytest.raises(ValueError) as excinfo:
+            obj.set_index(x="level_4")
+        assert str(excinfo.value) == "level_4 is not the name of an existing variable."
+
     def test_reset_index(self):
         indexes = [self.mindex.get_level_values(n) for n in self.mindex.names]
         coords = {idx.name: ("x", idx) for idx in indexes}
@@ -3139,6 +3148,56 @@ class TestDataArray:
         x2, = align(x, copy=True)
         assert_identical(x, x2)
         assert source_ndarray(x2.data) is not source_ndarray(x.data)
+
+    def test_align_override(self):
+        left = DataArray([1, 2, 3], dims="x", coords={"x": [0, 1, 2]})
+        right = DataArray(
+            np.arange(9).reshape((3, 3)),
+            dims=["x", "y"],
+            coords={"x": [0.1, 1.1, 2.1], "y": [1, 2, 3]},
+        )
+
+        expected_right = DataArray(
+            np.arange(9).reshape(3, 3),
+            dims=["x", "y"],
+            coords={"x": [0, 1, 2], "y": [1, 2, 3]},
+        )
+
+        new_left, new_right = align(left, right, join="override")
+        assert_identical(left, new_left)
+        assert_identical(new_right, expected_right)
+
+        new_left, new_right = align(left, right, exclude="x", join="override")
+        assert_identical(left, new_left)
+        assert_identical(right, new_right)
+
+        new_left, new_right = xr.align(
+            left.isel(x=0, drop=True), right, exclude="x", join="override"
+        )
+        assert_identical(left.isel(x=0, drop=True), new_left)
+        assert_identical(right, new_right)
+
+        with raises_regex(ValueError, "Indexes along dimension 'x' don't have"):
+            align(left.isel(x=0).expand_dims("x"), right, join="override")
+
+    @pytest.mark.parametrize(
+        "darrays",
+        [
+            [
+                DataArray(0),
+                DataArray([1], [("x", [1])]),
+                DataArray([2, 3], [("x", [2, 3])]),
+            ],
+            [
+                DataArray([2, 3], [("x", [2, 3])]),
+                DataArray([1], [("x", [1])]),
+                DataArray(0),
+            ],
+        ],
+    )
+    def test_align_override_error(self, darrays):
+        with raises_regex(ValueError, "Indexes along dimension 'x' don't have"):
+            xr.align(*darrays, join="override")
 
     def test_align_exclude(self):
         x = DataArray([[1, 2], [3, 4]], coords=[("a", [-1, -2]), ("b", [3, 4])])